requirements.txt
streamlit
google-generativeai
pandas
openpyxl
pillow
